Extract `waitUntil` into shared test helpers and fix silent timeout fail
Prompt:
In the splint repo, the `waitUntil` test helper is duplicated across multiple files and silently swallows timeout failures. When the deadline is exceeded without the condition becoming true, the function returns with no assertion — tests then proceed with incomplete setup and fail downstream for the wrong reasons.

Three files were already fixed as part of the GroupedLens work (see recent commits on this branch):
- [Tests/SplintTests/GroupedLensTests.swift](Tests/SplintTests/GroupedLensTests.swift)
- [Tests/SplintTests/GroupedLensLargeNTests.swift](Tests/SplintTests/GroupedLensLargeNTests.swift)
- [Example/BookshelfTests/ViewLogicTests.swift](Example/BookshelfTests/ViewLogicTests.swift)

The fix shape is:
```swift
private func waitUntil(
  timeout: Duration = .seconds(2),
  _ condition: () -> Bool,
  sourceLocation: SourceLocation = #_sourceLocation
) async {
  let deadline = ContinuousClock.now.advanced(by: timeout)
  while ContinuousClock.now < deadline {
    if condition() { return }
    try? await Task.sleep(for: .milliseconds(5))
  }
  #expect(condition(), "waitUntil timed out after \(timeout)", sourceLocation: sourceLocation)
}
```

Apply the same fix to every other test file that defines or duplicates this helper. Grep for `waitUntil` to find them — at minimum `Tests/SplintTests/LensTests.swift`, `Tests/SplintTests/LensLargeNTests.swift`, `Tests/SplintTests/CatalogTests.swift`, and anything under `Example/BookshelfTests/` that defines its own copy. Consider whether extracting to a shared test helper is worth it — probably yes for the SplintTests target, possibly separately for Example/BookshelfTests. Run `./script/test` and `./script/lint` to confirm clean.

Response:
I'll start by finding all `waitUntil` definitions and understanding the scope.I have the exact fix shape. The user asked me to consider extracting to shared helpers — I'll do that since 4 SplintTests files + 3 BookshelfTests files duplicate it. Let me also check the LensLargeN file's full helper signature (it has no comment) and verify naming clashes.Creating shared helpers and removing the per-file duplicates.Now remove the per-file private duplicates.Now run tests and lint.All 85 + 33 tests pass. Now lint.Lint error in `Credential.swift:149` is preexisting and unrelated. My changes lint clean. Signaling done.Extracted `waitUntil` into shared helpers and removed all 7 duplicated copies:

- New: [Tests/SplintTests/WaitUntil.swift](Tests/SplintTests/WaitUntil.swift), [Example/BookshelfTests/WaitUntil.swift](Example/BookshelfTests/WaitUntil.swift) — `@MainActor` top-level helper that emits `#expect(condition(), ...)` at the call site (via `#_sourceLocation`) when the deadline lapses.
- Removed private duplicates from: [LensTests](Tests/SplintTests/LensTests.swift), [LensLargeNTests](Tests/SplintTests/LensLargeNTests.swift), [CatalogTests](Tests/SplintTests/CatalogTests.swift), [JobTests](Tests/SplintTests/JobTests.swift), [ViewLogicTests](Example/BookshelfTests/ViewLogicTests.swift), [JobIntegrationTests](Example/BookshelfTests/JobIntegrationTests.swift), [CatalogIntegrationTests](Example/BookshelfTests/CatalogIntegrationTests.swift).

`./script/test`: 85 SplintTests + 33 BookshelfTests pass. `./script/lint`: only preexisting Credential.swift:149 issue remains (verified via stash — not introduced by this change).
