+++ Example/BookshelfTests/CatalogIntegrationTests.swift
@@ -6,17 +6,6 @@ import Splint
 @MainActor
 @Suite("Bookshelf catalog integration")
 struct CatalogIntegrationTests {
-  private func waitUntil(
-    timeout: Duration = .seconds(2),
-    _ condition: () -> Bool
-  ) async {
-    let deadline = ContinuousClock.now.advanced(by: timeout)
-    while ContinuousClock.now < deadline {
-      if condition() { return }
-      try? await Task.sleep(for: .milliseconds(5))
-    }
-  }
-
   @Test func loadPopulatesCatalog() async {
     let client = BookClient.mock
     let catalog = Catalog<Book, BookCriteria>(fetch: client.fetchBooks)
