+++ Example/BookshelfTests/ViewLogicTests.swift
@@ -17,19 +17,6 @@ struct ViewLogicTests {
     return c
   }
 
-  private func waitUntil(
-    timeout: Duration = .seconds(2),
-    _ condition: () -> Bool,
-    sourceLocation: SourceLocation = #_sourceLocation
-  ) async {
-    let deadline = ContinuousClock.now.advanced(by: timeout)
-    while ContinuousClock.now < deadline {
-      if condition() { return }
-      try? await Task.sleep(for: .milliseconds(5))
-    }
-    #expect(condition(), "waitUntil timed out after \(timeout)", sourceLocation: sourceLocation)
-  }
-
   // MARK: - displayLens: combined search + genre filter
 
   @Test func displayLensFiltersBySearchAndGenreTogether() async {
