+++ Tests/SplintTests/GroupedLensLargeNTests.swift
@@ -17,19 +17,6 @@ private final class LockCounter: @unchecked Sendable {
 struct GroupedLensLargeNTests {
   private let n = 1_000
 
-  private func waitUntil(
-    timeout: Duration = .seconds(2),
-    _ condition: () -> Bool,
-    sourceLocation: SourceLocation = #_sourceLocation
-  ) async {
-    let deadline = ContinuousClock.now.advanced(by: timeout)
-    while ContinuousClock.now < deadline {
-      if condition() { return }
-      try? await Task.sleep(for: .milliseconds(5))
-    }
-    #expect(condition(), "waitUntil timed out after \(timeout)", sourceLocation: sourceLocation)
-  }
-
   private func loadedCatalog(_ items: [TestItem]) async -> Catalog<TestItem, TestCriteria> {
     let c = Catalog<TestItem, TestCriteria> { _ in items }
     c.load(TestCriteria(category: "any"))
