+++ Tests/SplintTests/GroupedLensTests.swift
@@ -21,19 +21,6 @@ private enum Priority: String, Hashable, Comparable, Sendable {
 @MainActor
 @Suite("GroupedLens")
 struct GroupedLensTests {
-  private func waitUntil(
-    timeout: Duration = .seconds(2),
-    _ condition: () -> Bool,
-    sourceLocation: SourceLocation = #_sourceLocation
-  ) async {
-    let deadline = ContinuousClock.now.advanced(by: timeout)
-    while ContinuousClock.now < deadline {
-      if condition() { return }
-      try? await Task.sleep(for: .milliseconds(5))
-    }
-    #expect(condition(), "waitUntil timed out after \(timeout)", sourceLocation: sourceLocation)
-  }
-
   private let sample: [TestItem] = [
     TestItem(id: 1, name: "banana", score: 5),
     TestItem(id: 2, name: "apple", score: 9),
