+++ Tests/SplintTests/LensTests.swift
@@ -6,17 +6,6 @@ import Testing
 @MainActor
 @Suite("Lens")
 struct LensTests {
-  private func waitUntil(
-    timeout: Duration = .seconds(2),
-    _ condition: () -> Bool
-  ) async {
-    let deadline = ContinuousClock.now.advanced(by: timeout)
-    while ContinuousClock.now < deadline {
-      if condition() { return }
-      try? await Task.sleep(for: .milliseconds(5))
-    }
-  }
-
   private let sample: [TestItem] = [
     TestItem(id: 1, name: "banana", score: 5),
     TestItem(id: 2, name: "apple", score: 9),
